Use swift-dependencies for `AttachmentProvider`, refactor `AnnotationPersistenceManager` to use `@Dependency` macro for `AttachmentProvider`, remove attachmentProvider parameter from initializer
