+++ TrekPoint/Views/MapFeatureNavigator.swift
@@ -154,7 +154,7 @@ struct MapFeatureNavigator: View {
         selection: .constant(nil),
         annotations: [ModelContainer.previewAnnotation],
         polylines: [ModelContainer.previewPolyline],
-        annotationManager: AnnotationPersistenceManager(modelContainer: .preview, attachmentStore: AttachmentStore()),
+        annotationManager: AnnotationPersistenceManager(modelContainer: .preview),
         polylineManager: PolylinePersistenceManager(modelContainer: .preview),
         toastManager: .init(),
         onSelection: { _ in }
